refactor HTTP import tests to use registry constants and add importer pod annotation verification
