+++ tests/storage/cdi_import/test_import_registry.py
@@ -3,6 +3,7 @@ import logging
 import pytest
 from kubernetes.client.rest import ApiException
 from ocp_resources.datavolume import DataVolume
+from referencing import Resource
 
 from tests.storage.constants import QUAY_FEDORA_CONTAINER_IMAGE
 from tests.storage.utils import (
@@ -182,3 +183,21 @@ def test_public_registry_data_volume_archive(unprivileged_client, namespace, sto
             storage_class=[*storage_class_name_scope_function][0],
         ):
             return
+
+
+@pytest.mark.polarion("CNV-5509")
+@pytest.mark.s390x
+def test_importer_pod_annotation(dv_with_annotation, linux_nad):
+    """Verify "k8s.v1.cni.cncf.io/networks" can be passed to the importer pod"""
+    networks_annotation = f"{Resource.ApiGroup.K8S_V1_CNI_CNCF_IO}/networks"
+    network_status_annotation = f"{Resource.ApiGroup.K8S_V1_CNI_CNCF_IO}/network-status"
+
+    networks_value = dv_with_annotation.get(networks_annotation)
+    assert networks_value == linux_nad.name, (
+        f"DV annotation is not passed to the importer pod. Expected: {linux_nad.name}, Found: {networks_value}"
+    )
+
+    network_status_value = dv_with_annotation.get(network_status_annotation)
+    assert '"interface": "net1"' in network_status_value, (
+        f"DV annotation is not passed to the importer pod. Expected interface: net1, Found: {network_status_value}"
+    )
